iio: altera_adxcvr: Add skip-pll-reconfiguration option
Reconfiguration of the PLLs (ATX, CDR) is not mandatory after the FPGA
is programmed. The only required steps are related to their calibration.

Having this option will help identify the Quartus configuration
mismatches or bring up PLLs for which the reconfiguration procedure is
not documented or has not yet been implemented in the driver.

Signed-off-by: Dragos Bogdan <dragos.bogdan@analog.com>