iio: adc: ad7192: modify shorted channel
The shorted channel had the same id as the channel with address
AD7192_CH_AIN2 ("in_voltage2"). This caused an incompatibility with
libiio. The library requires all channels to have different ids. By
creating it with the differential channel macro, the shorted channel
gets an unique id ("in_voltage2-voltage2").

Signed-off-by: Alisa Roman <alisa.roman@analog.com>